+++ src/algorithms/signal_source/gnuradio_blocks/unpack_2bit_samples.cc
@@ -12,18 +12,7 @@
  *
  * This file is part of GNSS-SDR.
  *
- * GNSS-SDR is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * GNSS-SDR is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with GNSS-SDR. If not, see <https://www.gnu.org/licenses/>.
+ * SPDX-License-Identifier: GPL-3.0-or-later
  *
  * -------------------------------------------------------------------------
  */
@@ -164,7 +153,7 @@ int unpack_2bit_samples::work(int noutput_items,
         {
             if (swap_endian_bytes_)
                 {
-                    for (unsigned int i = 0; i < ninput_bytes; ++i)
+                    for (size_t i = 0; i < ninput_bytes; ++i)
                         {
                             // Read packed input sample (1 byte = 4 samples)
                             raw_byte.byte = in[i];
@@ -177,7 +166,7 @@ int unpack_2bit_samples::work(int noutput_items,
                 }
             else
                 {
-                    for (unsigned int i = 0; i < ninput_bytes; ++i)
+                    for (size_t i = 0; i < ninput_bytes; ++i)
                         {
                             // Read packed input sample (1 byte = 4 samples)
                             raw_byte.byte = in[i];
@@ -193,7 +182,7 @@ int unpack_2bit_samples::work(int noutput_items,
         {
             if (swap_endian_bytes_)
                 {
-                    for (unsigned int i = 0; i < ninput_bytes; ++i)
+                    for (size_t i = 0; i < ninput_bytes; ++i)
                         {
                             // Read packed input sample (1 byte = 4 samples)
                             raw_byte.byte = in[i];
@@ -206,7 +195,7 @@ int unpack_2bit_samples::work(int noutput_items,
                 }
             else
                 {
-                    for (unsigned int i = 0; i < ninput_bytes; ++i)
+                    for (size_t i = 0; i < ninput_bytes; ++i)
                         {
                             // Read packed input sample (1 byte = 4 samples)
                             raw_byte.byte = in[i];
